feat : 주문페이지 하단 프래그먼트 구현

+++ app/src/main/java/com/example/popmate/model/data/remote/order/StoreItem.kt
@@ -3,7 +3,7 @@ package com.example.popmate.model.data.remote.order
 import java.util.Date
 
 data class StoreItem(
-    val tbItemId: Long,
+    val tbItemId: Int,
     val storeId: Long,
     val name: String,
     val price: Int,
